+++ src/plugins/chatview/tests/message-audio.js
@@ -8,7 +8,7 @@ describe("A Chat Message", function () {
             mock.initConverse(['chatBoxesFetched'],
             { fetch_url_headers: true },
             async function (_converse) {
-        await mock.waitForRoster(_converse, 'current');
+        await mock.waitForRoster(_converse, 'current', 1);
         const base_url = 'https://conversejs.org';
         const message = base_url+"/logo/audio.mp3";
 
@@ -35,7 +35,7 @@ describe("A Chat Message", function () {
             });
         });
 
-        await mock.waitForRoster(_converse, 'current');
+        await mock.waitForRoster(_converse, 'current', 1);
         const message = 'http://foo.bar/stream';
         const contact_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
         await mock.openChatBoxFor(_converse, contact_jid);
@@ -51,7 +51,7 @@ describe("A Chat Message", function () {
             { fetch_url_headers: true },
             async function (_converse) {
 
-        await mock.waitForRoster(_converse, 'current');
+        await mock.waitForRoster(_converse, 'current', 1);
         const message = 'https://differentdrumz.radioca.st/stream/1/';
 
         const contact_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
@@ -63,5 +63,22 @@ describe("A Chat Message", function () {
         expect(msg.innerHTML.replace(/<!-.*?->/g, '').replace(/(\r\n|\n|\r)/gm, "").trim()).toEqual(
             `<audio controls="" src="${message}"></audio>`+
             `<a target="_blank" rel="noopener" href="${message}">${message}</a>`);
+    }));
+
+    it("will render Spotify player for Spotify URLs",
+            mock.initConverse(['chatBoxesFetched'],
+            { embed_3rd_party_media_players: true, view_mode: 'fullscreen' },
+            async function (_converse) {
+
+        await mock.waitForRoster(_converse, 'current', 1);
+        const message = 'https://open.spotify.com/track/6rqhFgbbKwnb9MLmUQDhG6';
+
+        const contact_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
+        await mock.openChatBoxFor(_converse, contact_jid);
+        const view = _converse.chatboxviews.get(contact_jid);
+        await mock.sendMessage(view, message);
+        await u.waitUntil(() => view.querySelectorAll('.chat-content iframe').length, 1000)
+        const msg = sizzle('.chat-content .chat-msg:last .chat-msg__text').pop();
+        expect(msg.querySelector('iframe').src).toContain('https://open.spotify.com/embed/track/6rqhFgbbKwnb9MLmUQDhG6');
     }));
 });